Lab1: Add client code and modify server code
The server first gives client the list of images
that it can send. After the client choose and send
the request to the server, it will process the image
and send it in batch also with the image size to
identify when the transfering process is complete.
After that, the client can choose which file to receive
again.